Delete table of contents
I never really liked that the generated TOC was part of the source.
There's probably a way to get it in code and, say, prepend it to the
readme but Org provides TOC-generation on export and exporting has other
nice features, and maybe I don't want to rely on Github nicely
formatting org files, so consider those options instead.