Fonts revamp (part 1): un-parameterize font-serif and font-unicode
They were only being used for Doom. I originally thought that setting
fonts explicitly would make Doom's behavior more predictible but a bunch
of stuff isn't working and I'm thinking both that a.) letting the
program pick the font might be more robust, and b.) that in the case of
unicode I don't care much which font is being used anyway.

The doom-serif-font is only used to se an Emacs face called
`fixed-pitch-serif` which as far as I can tell isn't really used. Like,
the docs say it's a "standard font" but none of my packages or themes use it.