Get rid of 'dbus?' flag
I recall doing this originally because I thought GDM started dbus
(dbus-run-session) whenever a user logged in but now I'm not so sure? In
any case the better option would be to make it not do that and keep
things more decomposed. Also not all login managers necessarily do
dbus-run-session (I would guess) so still don't assume it.